# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FIFA World Cup Stadium AI Traffic Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TicketInput {
  ticketId: string;
  userId: string;
}

export interface ScanInput {
  ticketId: string;
  gate: string;
}

export type ScaleInputAction = typeof ScaleInputAction[keyof typeof ScaleInputAction];


export const ScaleInputAction = {
  'add-server': 'add-server',
  'remove-server': 'remove-server',
} as const;

export interface ScaleInput {
  action: ScaleInputAction;
}

export type SimulationConfigIntensity = typeof SimulationConfigIntensity[keyof typeof SimulationConfigIntensity];


export const SimulationConfigIntensity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  surge: 'surge',
} as const;

export interface SimulationConfig {
  intensity?: SimulationConfigIntensity;
  durationSeconds?: number;
}

export interface SystemMetrics {
  avgLatency: number;
  cpuUsage: number;
  memoryUsage: number;
  activeServers: number;
  requestsPerSecond: number;
  errorRate: number;
  totalRequests?: number;
}

export interface TicketValidationResult {
  valid: boolean;
  processingTime: number;
  metrics?: SystemMetrics;
  /** @nullable */
  error?: string | null;
}

export interface ScanResult {
  success: boolean;
  gate: string;
  entryTime: string;
  totalEntered?: number;
}

export type GateStatusStatus = typeof GateStatusStatus[keyof typeof GateStatusStatus];


export const GateStatusStatus = {
  open: 'open',
  congested: 'congested',
  closed: 'closed',
} as const;

export interface GateStatus {
  id: string;
  name: string;
  status: GateStatusStatus;
  throughput: number;
}

export interface StadiumCapacity {
  totalCapacity: number;
  currentOccupancy: number;
  occupancyPercent: number;
  gates: GateStatus[];
}

export interface MetricsSnapshot {
  timestamp: number;
  avgLatency: number;
  cpuUsage: number;
  memoryUsage: number;
  requestsPerSecond: number;
  activeServers: number;
  errorRate?: number;
}

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  info: 'info',
  warning: 'warning',
  critical: 'critical',
} as const;

export interface Alert {
  id: string;
  severity: AlertSeverity;
  title: string;
  message: string;
  timestamp: number;
  resolved?: boolean;
  /** @nullable */
  aiAction?: string | null;
}

export interface ScaleResult {
  success: boolean;
  activeServers: number;
  metrics?: SystemMetrics;
}

export interface ResetResult {
  status: string;
}

export interface AiAnalysisResult {
  status: string;
  analysis: string;
  actions: string[];
  confidence: number;
  serversAdded?: number;
  metricsAfter?: SystemMetrics;
}

export interface SimulationStatus {
  running: boolean;
  stage: string;
  /** @nullable */
  intensity: string | null;
  elapsedSeconds: number;
  virtualUsers: number;
  /** @nullable */
  nextStage?: string | null;
}

export interface OverloadError {
  error: string;
  metrics?: SystemMetrics;
}