# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FIFA World Cup Stadium AI Traffic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ticket/validate:
    post:
      operationId: validateTicket
      tags: [tickets]
      summary: Validate a ticket QR code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "200":
          description: Ticket valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketValidationResult"
        "401":
          description: Invalid ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketValidationResult"
        "503":
          description: Service overloaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverloadError"
  /ticket/scan:
    post:
      operationId: scanTicket
      tags: [tickets]
      summary: Process stadium entry scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "200":
          description: Entry processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
...
  /admin/scale:
    post:
      operationId: scaleServer
      tags: [admin]
      summary: Manually scale server capacity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScaleInput"
      responses:
        "200":
          description: Scale action applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScaleResult"
...
  /simulation/start:
    post:
      operationId: startSimulation
      tags: [simulation]
      summary: Start a crowd surge simulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationConfig"
      responses:
        "200":
          description: Simulation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationStatus"
...
components:
  schemas:
...
    TicketInput:
      type: object
      required: [ticketId, userId]
      properties:
        ticketId:
          type: string
        userId:
          type: string
    ScanInput:
      type: object
      required: [ticketId, gate]
      properties:
        ticketId:
          type: string
        gate:
          type: string
    ScaleInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [add-server, remove-server]
    SimulationConfig:
      type: object
      properties:
        intensity:
          type: string
          enum: [low, medium, high, surge]
        durationSeconds:
          type: integer
    TicketValidationResult:
      type: object
      required: [valid, processingTime]
      properties:
        valid:
          type: boolean
        processingTime:
          type: integer
        metrics:
          $ref: "#/components/schemas/SystemMetrics"
        error:
          type: ["string", "null"]
    ScanResult:
      type: object
      required: [success, gate, entryTime]
      properties:
        success:
          type: boolean
        gate:
          type: string
        entryTime:
          type: string
        totalEntered:
          type: integer
...
    SystemMetrics:
      type: object
      required: [avgLatency, cpuUsage, memoryUsage, activeServers, requestsPerSecond, errorRate]
      properties:
        avgLatency:
          type: number
        cpuUsage:
          type: number
        memoryUsage:
          type: number
        activeServers:
          type: integer
        requestsPerSecond:
          type: number
        errorRate:
          type: number
        totalRequests:
          type: integer
...
    ScaleResult:
      type: object
      required: [success, activeServers]
      properties:
        success:
          type: boolean
        activeServers:
          type: integer
        metrics:
          $ref: "#/components/schemas/SystemMetrics"
...
    SimulationStatus:
      type: object
      required: [running, stage, intensity, elapsedSeconds, virtualUsers]
      properties:
        running:
          type: boolean
        stage:
          type: string
        intensity:
          type: ["string", "null"]
        elapsedSeconds:
          type: integer
        virtualUsers:
          type: integer
        nextStage:
          type: ["string", "null"]
    OverloadError:
      type: object
      required: [error]
      properties:
        error:
          type: string
        metrics:
          $ref: "#/components/schemas/SystemMetrics"