# lib/api-zod/src/generated/types/overloadError.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FIFA World Cup Stadium AI Traffic Management API
 * OpenAPI spec version: 0.1.0
 */
import type { SystemMetrics } from './systemMetrics';

export interface OverloadError {
  error: string;
  metrics?: SystemMetrics;
}