# lib/api-zod/src/generated/types/simulationStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FIFA World Cup Stadium AI Traffic Management API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationStatus {
  running: boolean;
  stage: string;
  /** @nullable */
  intensity: string | null;
  elapsedSeconds: number;
  virtualUsers: number;
  /** @nullable */
  nextStage?: string | null;
}