# lib/api-zod/src/generated/types/systemMetrics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FIFA World Cup Stadium AI Traffic Management API
 * OpenAPI spec version: 0.1.0
 */

export interface SystemMetrics {
  avgLatency: number;
  cpuUsage: number;
  memoryUsage: number;
  activeServers: number;
  requestsPerSecond: number;
  errorRate: number;
  totalRequests?: number;
}